insert new price evaluation on cron job

+++ internal/pkg/models/price_snapshot.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type PriceSnapshot struct {
-	Id          int64     `db:"price_snapshot_id"`
+	Id          int       `db:"price_snapshot_id"`
 	CreatedTime time.Time `db:"created_time"`
 	Price       float64   `db:"price"`
 }
